requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
idna==2.9
importlib-metadata==1.6.0
joblib==0.14.1
kiwisolver==1.2.0
matplotlib==3.2.1
more-itertools==8.2.0
nltk==3.5
numpy==1.18.4
packaging==20.3
pdfminer3k==1.3.1
Pillow==7.1.2
pluggy==0.13.1
ply==3.11
py==1.8.1
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==5.4.1
python-dateutil==2.8.1
regex==2020.4.4
requests==2.23.0
six==1.14.0
slate3k==0.5.3
tika==1.24
tqdm==4.46.0
urllib3==1.25.9
wcwidth==0.1.9
wordcloud==1.7.0
zipp==3.1.0
